--- src/home.js
@@ -1,11 +1,0 @@
-
-// let open_sideBar = document.getElementById("button-openSideBar").addEventListener("click", () => {
-//     document.getElementById("mySidebar").style.display = "block";
-// });
-// let close_sideBar = document.getElementById("button-closeSideBar").addEventListener("click", () => {
-//     document.getElementById("mySidebar").style.display = "none";
-// });
-
-require("./model/SofkianFunctions.js");
-require("./model/SofkianClases.js");
-require("./HtmlController");
